j'avance

+++ db/migrate/20140107174837_add_details_to_articles.rb
@@ -1,6 +1,6 @@
 class AddDetailsToArticles < ActiveRecord::Migration
   def change
     add_column :articles, :part_number, :string
-    add_column :articles, :price, :decimal
+    add_column :articles, :price, :integer
   end
 end
